moved intreface files from types dir to interfaces dir and organised urls for reading frameworks
