+++ src/components/SearchBar.tsx
@@ -2,7 +2,7 @@ import React from 'react';
 import TextField from '@mui/material/TextField';
 import InputAdornment from '@mui/material/InputAdornment';
 import SearchIcon from '@mui/icons-material/Search';
-import { SearchBarProps } from '@/types/BaseInterface';
+import { SearchBarProps } from '@/interfaces/BaseInterface';
 
 const SearchBar: React.FC<SearchBarProps> = ({
   value,
