+++ src/interfaces/BaseInterface.ts
