+++ src/interfaces/CategoryInterface.ts
