+++ src/interfaces/ChannelInterface.ts
