+++ src/interfaces/FrameworkInterface.ts
