+++ src/interfaces/TermInterface.ts
