+++ src/utils/url.config.ts
@@ -2,6 +2,7 @@ export const URL_CONFIG = {
   API: {
     COMPOSITE_SEARCH: `${process.env.NEXT_PUBLIC_INTERFACE_URL}/action/composite/v3/search`,
     CHANNEL_CREATE: `${process.env.NEXT_PUBLIC_INTERFACE_URL}/action/channel/v3/create`,
+    FRAMEWORK_READ: `${process.env.NEXT_PUBLIC_INTERFACE_URL}/api/framework/v1/read`,
     // Add more API endpoints here as needed
   },
 };
